Added informational text
There's now a small black bar at the bottom of the window which displays the current file's path, the size of the image being displayed relative to the image's original size and the file index within the directory. The option to change the font size is also added to the config file.