Multithreading & refactoring
Noticed that the usability could be improved using multithreading to offload image loading onto another thread. This required more effort than I first estimated and led to several findings:

1) OpenGL contexts are very thread-local, this means that the thread which initializes a context must be the one to use it. This meant that the `Image` class could not use the renderer to create a texture when being constructed by the worker thread. The main thread (or a dedicated rendering thread if any) must be notified and carry out the operation instead because it owns the context.

2) Event handling on most platforms (e.g. Windows) is also thread-local. As SDL calls native WinAPI handles under the hood it requires that the main thread be the one that recieves and handles events. So while the `SDL_EventFilter` may run on a different thread (but does not on Windows as I found) the call to `SDL_PumpEvents()` must be made on the main thread.

The `Image` class alone dealing in terms of `SDL_Surface`s was not elegant and thus `Widget`s were also made to also work this way. The `Drawable` class is used by the `draw()` method for (effectively) converting surfaces -that are modified by the derived class' methods- to textures. A pair of `Semaphore` and `B(inary)Semaphore` implementations were also included (can't wait for C++20).

In order to remove any doubt of race conditions a mutex is included with (almost) every class. It's most likely unnecessary as the logic of the program does not allow for concurrent modification; but better safe than sorry (with negligible performance impact).

Refactored the entire project in order to remove the `Control` singleton. Managed to do this while still ensuring encapsulation of all the control methods by simply not declaring them in `control.h`. Also reduced the amount of scroll when using up/down arrow keys to 2%.